- change initID as uint128_t
[ ] May break build
[ ] May break existing applications (see CHANGES.txt)
[ ] Bugfix
[ ] New Feature
[x] Cleanup
[x] Optimization
[ ] Documentation

+++ lib/client/channelPackets.h
@@ -34,7 +34,7 @@ namespace eq
                 size    = sizeof( ChannelConfigInitPacket );
             }
 
-        uint32_t        initID;
+        uint128_t        initID;
     };
 
     struct ChannelConfigInitReplyPacket : public ChannelPacket
